+++ addon/-task-state-mixin.js
@@ -34,6 +34,7 @@ export default Ember.Mixin.create({
   lastErrored:    alias('_scheduler.lastErrored'),
   lastCanceled:   alias('_scheduler.lastCanceled'),
   lastIncomplete: alias('_scheduler.lastIncomplete'),
+  performCount: alias('_scheduler.performCount'),
 
   numRunning: 0,
   numQueued: 0,
@@ -49,44 +50,5 @@ export default Ember.Mixin.create({
 
   _scheduler: null,
 
-  /* TODO: re-add this to work w task groups... right now it's coupled to .performs
-
-  _performs: computed('_performsPath', function() {
-    let path = this.get('_performsPath');
-    if (!path) { return; }
-
-    let task = this.context.get(path);
-    if (!(task instanceof Task)) {
-      throw new Error(`You wrote .performs('${path}'), but the object at '${path}' is not a Task`);
-    }
-    return task;
-  }),
-
-  _performsState: computed('_performs.nextPerformState', function() {
-    return this.get('_performs.nextPerformState') || 'succeed';
-  }),
-
-  nextPerformState: computed('_performsState', function() {
-    let performsState = this.get('_performsState');
-    return isSuccess(performsState) ?
-      this._bufferPolicy.getNextPerformStatus(this) :
-      performsState;
-  }),
-
-  performWillSucceed: computed('nextPerformState', function() {
-    return isSuccess(this.get('nextPerformState'));
-  }),
-
-  performWillDrop: computed.equal('nextPerformState', 'drop'),
-
-  performWillEnqueue: computed.equal('nextPerformState', 'enqueue'),
-
-  performWillCancelPrevious: computed.equal('nextPerformState', 'cancel_previous'),
-  */
-
 });
-
-//function isSuccess(nextPerformState) {
-  //return nextPerformState === 'succeed' || nextPerformState === 'cancel_previous';
-//}